Move generic aux-source composition out of detector_view_specs
The dynamic-transform aux-source plumbing is cross-cutting, not
detector-view-specific. Move it to its natural home so
`Instrument.register_spec` no longer has to reach upward into
`detector_view_specs`:

- `CombinedAuxSources` -> `config/workflow_spec.py` (next to
  `AuxSources`; pure composition primitive).
- `_DynamicTransformAuxSources` + `compose_aux_sources` ->
  `handlers/dynamic_transforms.py`. Fold the former
  `dynamic_transform_aux_inputs` / `dynamic_transform_routes`
  helpers into the class -- they were only used by it.

`detector_view_specs.py` retains `DetectorROIAuxSources`, which
genuinely is detector-specific.
